fix shellcheck

+++ main.sh
@@ -15,7 +15,8 @@ git config --global --add safe.directory "$tmp_dir"
 git clone "$INPUT_GITHUB_SERVER_URL/$INPUT_REPOSITORY.wiki.git" "$tmp_dir" --depth 1
 
 # Hidden files (like .myfile.txt, .git/, or .gitignore) are NOT copied.
-rm -rf "$tmp_dir"/*
+# The magic "${var:?}" makes it error if the var is zero-length/null.
+rm -rf "${tmp_dir:?}"/*
 cp -afv "$INPUT_PATH"/* "$tmp_dir"
 
 cd "$tmp_dir"
